Fix clippy::cognitive_complexity warnings
https://rust-lang.github.io/rust-clippy/master/index.html#cognitive_complexity

Verified with:

    cargo clippy --all-targets --all-features -- -A clippy::all -D clippy::cognitive_complexity

Note that this disables this lint for the one method, `stats::stats`,
and a handful of tests that triggers the warning. Fixing this was a
larger effort that I want to get into right now, so I opened #550 to
track coming back to clean this up for real.

FIXUP: cognitive_complexity