Remove unused STDistance calculation in conflate
When conflate queries for possible duplicate addresses it calculates the
distance from each possible duplicate to the feature being considered
using STDistance. However, this result is never used for anything as the
query selects all results within a radius of 0.01 degrees using
STWithin. Removing the unnecessary distance calculation speeds up the
query slightly, especially in the case when there are **many** possible
duplicates.